PullHistoryModel was added to DI and implemented in ViewModels, so both ViewModels can operate on same data. Added commands for 5* characters and Visibility switch(After losing 50/50 it's impossible to get standard character).

+++ GenshinWishCounter1.5/MVVM/Model/PullModel.cs
@@ -11,12 +11,12 @@ namespace GenshinWishCounter1._5.MVVM.Model
     /// </summary>
     public class PullModel
     {
-        public byte _pulledAt { get; set; }
+        public int _pulledAt { get; set; }
         public string _element { get; set; }
         public string _characterName { get; set; }
         public string _fiftyFiftyResult { get; set; }
 
-        public PullModel(byte pulledAt, string element, string characterName, string fiftyFiftyResult)
+        public PullModel(int pulledAt, string element, string characterName, string fiftyFiftyResult)
         {
             _pulledAt = pulledAt;
             _element = "/Images/" + element + ".png";
